+++ blockchain/sync_service.go
@@ -1,9 +1,17 @@
 package blockchain
 
 type PeerService interface {
-	GetRandomPeerId() PeerId
-	GetLastBlock(peerId PeerId) Block
-	GetBlockByHeight(peerId PeerId, height BlockHeight) Block
+	PeerInnerService
+	PeerOuterService
+}
+
+type PeerInnerService interface {
+	GetRandomPeer() Peer
+}
+
+type PeerOuterService interface {
+	GetLastBlock(peer Peer) Block
+	GetBlockByHeight(peer Peer, height BlockHeight) Block
 }
 
 type BlockService interface {
