+++ p2p/service.go
@@ -17,7 +17,7 @@ type PeerService interface {
 	DeliverPeerLeaderTable(connectionId string, peerLeaderTable PeerLeaderTable) error
 }
 
-type LeaderService interface{
+type LeaderService interface {
 	DeliverLeaderInfo(connectionId string, leader Leader) error
 }
 type Publish func(exchange string, topic string, data interface{}) (err error) // 나중에 의존성 주입을 해준다.
